Refactor: CSS 분리 - css 분리
