+++ src/styles/landingSlideCSS.ts
